+++ tests/testthat/test_define_variance_wrapper.R
@@ -6,7 +6,7 @@ test_that("variance_wrapper can be defined in globalenv()", {
   expect_error({
     variance_wrapper <<- define_variance_wrapper(
       variance_function = function(y) abs(colSums(y)), 
-      auxiliary_data = list(reference_id = ict_survey$firm_id),
+      reference_id = ict_survey$firm_id,
       default = list(id = "firm_id", weight = "w_calib", stat = "mean")
     )    
     variance_wrapper(ict_survey, speed_quanti)
@@ -19,10 +19,10 @@ test_that("variance_wrapper can be defined in another function", {
     preparation_function <- function(){
       a <- 1
       define_variance_wrapper(
-        variance_function = function(y) abs(colSums(y)) + 1, 
-        auxiliary_data = list(reference_id = ict_survey$firm_id),
-        default = list(id = "firm_id", weight = "w_calib", stat = "mean"), 
-        objects_to_include = "a"
+        variance_function = function(y, a) abs(colSums(y)) + a, 
+        reference_id = ict_survey$firm_id,
+        technical_data = list(a = a),
+        default = list(id = "firm_id", weight = "w_calib", stat = "mean")
       )
     }
     variance_wrapper2 <- preparation_function()
@@ -35,31 +35,19 @@ test_that("variance_wrapper can be defined in another function", {
 })
 
 test_that("variance_wrapper may use a reference_id specified as an unevaluated expression", {
-  expect_warning({
+  expect_error({
     id_list <- list(firm = ict_survey$firm_id)
     variance_wrapper <- define_variance_wrapper(
       variance_function = function(y, level = "firm") abs(colSums(y)), 
-      auxiliary_data = list(reference_id = ict_survey$firm_id, id_list = id_list),
+      reference_id = quote(id_list[[level]]),
       arg_type = list(data = "y", param = "level", aux = NULL),
-      default = list(id = "firm_id", weight = "w_calib", stat = "mean")
+      default = list(id = "firm_id", weight = "w_calib", stat = "mean"),
+      objects_to_include = "id_list"
     )
     rm(id_list)
     variance_wrapper(ict_survey, speed_quanti)
   }, regexp = NA)
 })
-
-test_that("variance_wrapper can be defined using reference_id backward compatibility", {
-  expect_warning({
-    variance_wrapper <<- define_variance_wrapper(
-      variance_function = function(y) abs(colSums(y)), 
-      reference_id = ict_survey$firm_id,
-      default = list(id = "firm_id", weight = "w_calib", stat = "mean")
-    )    
-    variance_wrapper(ict_survey, speed_quanti)
-  }, regexp = "reference_id is deprecated")
-})
-
-  
 
 test_that("variance_wrapper works in common situations", {
   expect_error(variance_wrapper(ict_survey, speed_quanti), regexp = NA)